* Refactored SearchBar into own component * fixed merge errors
