+++ app/src/components/ui/SearchBar.vue
@@ -1,25 +1,99 @@
 <template>
-    <v-autocomplete
-            filled
+    <v-combobox
+            style="background: white"
             rounded
             color="white"
-            label="z.B. MacherIn"
+            label="z.B. Macher/in"
             append-icon="search"
-            :items="volunteerTags">
-    </v-autocomplete>
+            item-text="tag"
+            autocomplete="off"
+            :items="mySearchProposals"
+            @input="addSearchTag"
+            :search-input.sync="currentSearchValue"
+    ></v-combobox>
 </template>
 
-<script>
-export default {
-    data: () => ({
-        volunteerTags: [
-            'MacherIn',
-            'DenkerIn',
-            'Jugend Arbeit',
-            'Soziales',
-        ]
-    })
-};
+<script lang="ts">
+import Vue from 'vue';
+import TagService from '@/utils/services/TagService';
+import Tag from '@/models/tag';
+import {mapActions, mapState} from 'vuex';
+
+export default Vue.extend({
+    data(): {
+        currentSearchValue: string
+    } {
+        return {
+            currentSearchValue: ''
+        };
+    },
+    created(): void {
+        this.initializeSearchProposals(TagService.getTags());
+    },
+    watch: {
+        selectedTag(newValue): void {
+            this.setSelectedTag(newValue);
+        },
+    },
+    computed: {
+        ...mapState(['searchProposals', 'selectedLocation', 'radiusSearchValue']),
+        mySearchProposals(): string[] {
+            if (!this.currentSearchValue || this.currentSearchValue.length < 2) {
+                return [];
+            }
+            const searchTerm = this.currentSearchValue;
+            const listOfMatchingTerms = this.matchSearchInput(searchTerm, this.searchProposals
+                    .filter((element) => 'label' in element));
+            const rankedListOfOrderedTerms = this.rankTerms(searchTerm, listOfMatchingTerms);
+
+            return rankedListOfOrderedTerms;
+        }
+    },
+    methods: {
+        ...mapActions(['initializeSearchProposals', 'setSelectedTag']),
+        matchSearchInput(searchTerm: string, proposals: Tag[]): string[] {
+            const stringArray: string[] = [];
+            searchTerm = searchTerm.toLowerCase();
+
+            proposals.forEach((tag) => {
+                if (tag.label.toLowerCase().match(searchTerm)) {
+                    stringArray.push(tag.label);
+                } else {
+                    tag.synonyms.forEach((element) => {
+                        if (element.toLowerCase().match(searchTerm)) {
+                            stringArray.push(tag.label + ' (' + element + ')');
+                        }
+                    });
+                }
+            });
+            return stringArray;
+        },
+        rankTerms(searchTerm: string, terms: string[]): string[] {
+            return terms.map((term) => {
+                // 2x on start; 1x on end, 0.5x in the middle
+                const rank = term.toLowerCase().startsWith(searchTerm.toLowerCase())
+                        ? 2
+                        : term.toLowerCase().endsWith(searchTerm.toLowerCase())
+                                ? 1
+                                : 0.5;
+                return {
+                    label: term,
+                    rank
+                };
+            })
+                    .sort((a, b) => Math.sign(b.rank - a.rank))
+                    .map((obj) => obj.label);
+        },
+        addSearchTag(tag: string): void {
+
+            const tagName = tag.includes(' (')
+                    ? tag.substr(0, tag.indexOf(' ('))
+                    : tag;
+
+            this.$emit('input', tagName);
+        },
+    }
+});
 </script>
 
 <style scoped>
