+++ app/src/views/Home.vue
@@ -20,18 +20,9 @@
       <v-form class="mt-12 mb-12">
         <v-row justify="center">
           <v-col cols="12" md="8">
-            <v-combobox
-              style="background: white"
-              rounded
-              color="white"
-              label="z.B. Macher/in"
-              append-icon="search"
-              item-text="tag"
-              autocomplete="off"
-              :items="mySearchProposals"
-              @input="addSearchTag"
-              :search-input.sync="currentSearchValue"
-            ></v-combobox>
+            <search-bar
+              @input="handleSearchEvent"
+            />
           </v-col>
         </v-row>
 
@@ -85,13 +76,13 @@ import 'vue-slick-carousel/dist/vue-slick-carousel-theme.css';
 import 'vue-slick-carousel/dist/vue-slick-carousel.css';
 import LocationSearchBar from '@/components/ui/LocationSearchBar.vue';
 import Radius from '@/components/ui/Radius.vue';
-import Post from '../models/post';
-import TagService from '../utils/services/TagService';
 import Tag from '@/models/tag';
+import SearchBar from '@/components/ui/SearchBar.vue';
 
 
 export default Vue.extend({
   components: {
+    SearchBar,
     VueSlickCarousel,
     LocationSearchBar,
     Radius,
@@ -99,7 +90,6 @@ export default Vue.extend({
   },
   data(): {
     volunteerTags: Array<{title: string, img: string}>,
-    selectedTag: string,
   } {
     return {
       volunteerTags: [
@@ -120,74 +110,21 @@ export default Vue.extend({
           img: require('../../public/images/sozial.jpeg')
         }
       ],
-      selectedTag: '',
     };
   },
   computed: {
-    ...mapState(['searchProposals', 'selectedLocation', 'radiusSearchValue']),
-    mySearchProposals(): string[] {
-      if (!this.currentSearchValue || this.currentSearchValue.length < 2) {
-        return [];
-      }
-      const searchTerm = this.currentSearchValue;
-      const listOfMatchingTerms = this.matchSearchInput(searchTerm, this.searchProposals
-        .filter((element) => 'label' in element));
-      const rankedListOfOrderedTerms = this.rankTerms(searchTerm, listOfMatchingTerms);
-
-      return rankedListOfOrderedTerms;
-    }
-  },
-  watch: {
-    selectedTag(newValue, oldValue): void {
-      this.setSelectedTag(newValue);
-    },
+    ...mapState(['selectedLocation', 'radiusSearchValue'])
   },
   methods: {
-    ...mapActions(['initializeSearchProposals', 'setSelectedTag']),
-    addSearchTag(tag: {tag: string}): void {
-      const tagName = tag.substr(0, tag.indexOf(' ('));
-
+    handleSearchEvent(searchValue: string) {
       this.$router.push({
         name: 'resultPage',
         query: {
-          q: tagName,
+          q: searchValue,
           city: this.selectedLocation,
           radius: this.radiusSearchValue
         }
       });
-    },
-    matchSearchInput(searchTerm: string, proposals: Tag[]): string[] {
-      const stringArray: string[] = [];
-      searchTerm = searchTerm.toLowerCase();
-
-      proposals.forEach((tag) => {
-        if (tag.label.toLowerCase().match(searchTerm)) {
-          stringArray.push(tag.label);
-        } else {
-          tag.synonyms.forEach((element) => {
-            if (element.toLowerCase().match(searchTerm)) {
-              stringArray.push(tag.label + ' (' + element + ')');
-            }
-          });
-        }
-      });
-      return stringArray;
-    },
-    rankTerms(searchTerm: string, terms: string[]): string[] {
-      return terms.map((term) => {
-          // 2x on start; 1x on end, 0.5x in the middle
-          const rank = term.toLowerCase().startsWith(searchTerm.toLowerCase())
-            ? 2
-            : term.toLowerCase().endsWith(searchTerm.toLowerCase())
-              ? 1
-              : 0.5;
-          return {
-            label: term,
-            rank
-          };
-        })
-        .sort((a, b) => Math.sign(b.rank - a.rank))
-        .map((obj) => obj.label);
     }
   }
 });
